Refactor how form submission works and img variable
I removed the onClick from the submit button and instead placed an
onSubmit event listener on the form element. This was done so that the
the required attribute on the textarea would function properly.

In addition to this refactor, I also reworked the img variable to be an
object so that it could contain both images respective src and alt
attributes. This was the best way to ensure unique alt text for both
images that I could think of.