Clean response data before assigning to useState
I created a utils file to house the cleaning function:
'cleanResponseData' This function prunes the response of all unnecessary
data, keeping only the id and the response text.

I then imported this cleaning function into the apiCalls file and then
invoked it with the response data from the post request.