Add state and pass it down to PromptForm
I created a new instance of state called responseLog. This will be used
to keep track of and add to the response log. The setResponseLog method
is being passed down to PromptForm via props and invoked in submitForm
so as to add the new prompt/response combo to the responseLog state.

Eventually this array will then be passed down to another component to
seed the data for the response log/cards.