Display AI response on the DOM
I created another instance of state called response to store the
response from the OpenAPI post request. To display it on the DOM I
declared a new variable, 'responseBox' which conditionally renders
when there is a response from the post request.

I also made the new DOM elements responsive to multiple viewports.