Refactor tests to account for network requests
A post request is now initiated on submission of the form. As such, I
went back and refactored one of the previous made tests to account for
this new functionality.

Now the test intercepts the post requests and stubs it with data I that
I created in the fixture file: 'examplePostResponse'. I also added
another test to make sure that the form cannot be submitted if no prompt
is provided.