Create components to display response log
I created the ResponseLog and ResponseCard components. The ResponseLog
component maps through the responseLog array that is passed to it as a
prop from App.js, and creates the a ResponseCard for each response in
the array. The ResponseCard component compromises the simply the HTML
that makes up the response cards.

For each I these components have accompanying css files that contain
their respective styling.